🌱 Split MissionDetailView into header/meta/tabs sub-components
MissionDetailView.tsx drops from 589 to 116 lines by extracting
MissionDetailHeader, MissionDetailMeta, MissionDetailTabNav and
MissionDetailTabContent into sibling files. Pure move — no markup,
logic, or exported-API changes.

Refs #21786
